Move activity actor definitions to separate header
