+++ src/game.cpp
@@ -33,6 +33,7 @@
 #include "achievement.h"
 #include "action.h"
 #include "activity_actor.h"
+#include "activity_actor_definitions.h"
 #include "activity_handlers.h"
 #include "artifact.h"
 #include "auto_note.h"
