+++ src/pickup.cpp
@@ -12,6 +12,7 @@
 #include <vector>
 
 #include "activity_actor.h"
+#include "activity_actor_definitions.h"
 #include "auto_pickup.h"
 #include "avatar.h"
 #include "cata_utility.h"
