+++ src/ranged.cpp
@@ -12,6 +12,7 @@
 #include <utility>
 #include <vector>
 
+#include "activity_actor_definitions.h"
 #include "avatar.h"
 #include "ballistics.h"
 #include "bodypart.h"
