+++ tests/invlet_test.cpp
@@ -8,6 +8,7 @@
 #include <utility>
 
 #include "activity_actor.h"
+#include "activity_actor_definitions.h"
 #include "avatar.h"
 #include "catch/catch.hpp"
 #include "inventory.h"
